refactor: report 관련 상수들 별도의 파일로 분리
